% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_PhenoFlex_GDHwrapper_v2.R
\name{custom_PhenoFlex_GDHwrapper_v2}
\alias{custom_PhenoFlex_GDHwrapper_v2}
\title{Function to return bloom dates, can also return accumulated chill and heat}
\usage{
custom_PhenoFlex_GDHwrapper_v2(x, par, constraints = FALSE)
}
\arguments{
\item{x}{data.frame with at least the column "Temp" (hourly temperature data) and "JDay" (day of the year)}

\item{par}{vector of length 12 with the PhenoFlex model parameters in the following order:
yc, zc, s1, Tu, E0, E1, A0, A1, Tf, Tc, Tb, slope. If the set of parameters include 
theta_star, theta_c, tau and pie_c instead of E0, E1, A0, and A1, then the function \code{\link{convert_parameters}}
should be run before inserting the parameters}

\item{constraints}{boolean, set FALSE by default. If set TRUE, the function will perform
some checks on the model parameter prior to the prediction of bloom day. For more information 
see "Details"}
}
\value{
single numeric value with the day of the year, for which the model predicts bloom
with given temperature data and model parameters.
}
\description{
Takes hourly weather time series and PhenoFlex model parameters and returns day of the year with bloom.
}
\details{
This function is mainly used by the global optimization objective functions
(\code{\link[evalpheno]{evaluation_function_meigo}, \link[evalpheno]{evaluation_function_meigo_nonlinear}, \link[evalpheno]{evaluation_function_meigo_vns}}) 
to fit the PhenoFlex parameters. For a single season hourly temperature data time series and a set of model
parameters it returns the predicted day of the year of bloom.

In case constraints is TRUE, logic checks will be performed for some parameters. If one of the checks 
is violated, then the function will return "NA" instead of the predicted bloom day. Checks involve parameters
of the dynamic chill model and the growing degree hour (GDH) model. Tests include:
\itemize{
 \item{Parameters of GDH model:}
    \itemize{
    \item{Tf > Tb}
    \item{Tc > Tb}
    }
 \item{Dynamic model}
 \itemize{
    \item{exp((10 * E0)/(297 * 279)) > 1.5 & exp((10 * E0)/(297 * 279)) < 3.5}
    \item{exp((10 * E1)/(297 * 279)) > 1.5 | exp((10 * E0)/(297 * 279)) < 3.5}
    }
}

The parameter test of the dynamic model follows the approach suggested by Egea et al (2021)
in which the Q10 parameter was used to constraint the E0 and E1 parameters. The Q10
parameter how much a chemical process increases by increasing the temperature by 10 degree 
Celsius. Biological reactions typically a Q10 value ranging from 2 to 3. Here, ranges of 1.5 to 3.5 were used as a threshold.
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
