% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_all_daoptim.R
\name{eval_all_daoptim}
\alias{eval_all_daoptim}
\title{Evaluation funciton full model, general format
 
Takes model parameters from PhenoFlex, Sequential Model, Parallel Model or Partial Overlap Model}
\usage{
eval_all_daoptim(
  x,
  modelfn,
  bloomJDays,
  SeasonList,
  na_penalty = 365,
  intermed_chill = FALSE,
  pos_int = c(5:8),
  check_constrain = FALSE,
  pos_E0_E1 = c(5, 6),
  pos_heat = c(11, 4, 10),
  c_L = c(0, 0, 0, 1.5, 1.5),
  c_U = c(Inf, Inf, Inf, 3.5, 3.5)
)
}
\arguments{
\item{x}{model parameters needs to fit the format of PhenoFlex, Parallel Model, Sequential Model or Partial Overlap Model}

\item{modelfn}{function used within the evaluation function to calculate the actual bloomday}

\item{bloomJDays}{numeric containing the days of the year with the observed bloom}

\item{SeasonList}{list of hourly temperatures for the individual phonological seasons. Each element should contain a data.frame
with the columns "Temp" (for the hourly temperature) and "JDay" for the corresponding Julian day. Is usually
generated using \link[chillR]{genSeasonList}}

\item{na_penalty}{numeric, by default 365. Penalty for the phenology
prediction function when it fails return a bloom prediction}

\item{intermed_chill}{boolean, by default `FALSE`. If `TRUE` converts intermediate parameters of chill submodel (theta_star, theta_c, pie_c, tau) to E0, E1, A0, A1. 
Intermediate parameters have tighter search space and optimization algorithms find faster combinations.}

\item{pos_int}{numeric vector of length 4. Tells the position of intermediate parameters in vector `x` that should be converted 
if `intermed_chill=TRUE`. Assumes order: theta_star, theta_c, piec and tau.}

\item{check_constrain}{booloean, by default `FALSE`. If set `TRUE` additional constraints will be evaluated. This includes
Q10 metric for parameters E0 and E1 should be in plausible range for biological processes (between 1.5 and 3.5) and 
logical constraints in GDH model (Tb <= Tu <= Tc). If constraints are violated, function does not compute prdiction error
but Inf}

\item{pos_E0_E1}{numeric vector of length 2, indicates position of E0 and E1 in the parameter vector `x`. Only relevant when checking for constraints.}

\item{pos_heat}{numeric vector of length 3, indicates position of Tb, Tu and Tc in the parameter vector `x`. Only relevant when checking for constraints.}

\item{c_L}{numeric vector of length 5, lower tolerance level of constraints.}

\item{c_U}{numeric vector of length 5, upper tolerance level of constraints.}
}
\value{
numeric, sum of squared difference between predicted and observed phenology observations.
}
\description{
The function is parsed to the optimization algorithm during calibration. Format is compatible with
GenSA and DEoptim optimizers. Allows for additional parameter constraints. Also can allow
parameter conversion for chill submodel (Dynamic Model). Returns the squared sum of prediction errors or
Inf if constraints are violated.
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
