% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_fixed_daoptim.R
\name{eval_fixed_daoptim}
\alias{eval_fixed_daoptim}
\title{Evaluation funciton full model, general format
 
Takes model parameters from PhenoFlex, Sequential Model, Parallel Model or Partial Overlap Model}
\usage{
eval_fixed_daoptim(
  x,
  modelfn,
  bloomJDays,
  SeasonList,
  na_penalty = 365,
  E0 = 4153.5,
  E1 = 12888.8,
  A0 = 139500,
  A1 = 2.567e+18,
  Tf = 4,
  slope = 1.6,
  Tb = 4,
  Tu = 25,
  Tc = 36
)
}
\arguments{
\item{x}{model parameters needs to fit the format of PhenoFlex, Parallel Model, Sequential Model or Partial Overlap Model
Assumes only model specific parameters, excluding parameters of the submodel for chill and heat accumulation.}

\item{modelfn}{function used within the evaluation function to calculate the actual bloomday}

\item{bloomJDays}{numeric containing the days of the year with the observed bloom}

\item{SeasonList}{list of hourly temperatures for the individual phonological seasons. Each element should contain a data.frame
with the columns "Temp" (for the hourly temperature) and "JDay" for the corresponding Julian day. Is usually
generated using \link[chillR]{genSeasonList}}

\item{na_penalty}{numeric, by default 365. Penalty for the phenology
prediction function when it fails return a bloom prediction}

\item{E0}{numeric. Parameter \eqn{E_0}{E0} of the dynamic model}

\item{E1}{numeric. Parameter \eqn{E_1}{E1} of the dynamic model}

\item{A0}{numeric. Parameter \eqn{A_0}{A0} of the dynamic model}

\item{A1}{numeric. Parameter \eqn{A_1}{A1} of the dynamic model}

\item{Tf}{numeric. Transition temperature (in degree Kelvin) for the sigmoidal function}

\item{slope}{numeric. Slope parameter for sigmoidal function}

\item{Tb}{numeric. GDH base temperature (lower threshold)}

\item{Tu}{numeric. GDH optimal temperature}

\item{Tc}{numeric. GDH upper temperature (upper threshold)}
}
\value{
numeric, sum of squared difference between predicted and observed phenology observations.
}
\description{
The function is parsed to the optimization algorithm during calibration. Format is compatible with
GenSA and DEoptim optimizers. Allows for additional parameter constraints. Also can allow
parameter conversion for chill submodel (Dynamic Model). Returns the squared sum of prediction errors or
Inf if constraints are violated.
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
