% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDH_wrapper.R
\name{eval_function_gdh}
\alias{eval_function_gdh}
\title{Evaluation function for ecodormancy}
\usage{
eval_function_gdh(x, SeasonList, dispersion_fun = "calc_cv")
}
\arguments{
\item{x}{model parameters of GDH model, order: Tb, Tu, Tc}

\item{SeasonList}{List of SeasonLists. For each cultivar one entry. Individual seasonlists 
per cultivar contain exactly period of: measured end of endodormancy release to measured
ecodormancy release.}

\item{dispersion_fun}{name of function to measure dispersion of accumulated GDH per
cultivar. Needs to be a function loaded in the environment. Can also be a custom function.
By default "calc_cv", which calculates the coefficient of variation.}
}
\value{
sum of the dispersion function. This will be subject to global optimization.
}
\description{
Takes model parameters for GDH model and list of seasons with daily temperature time series.
}
\details{
The function evaluates a set of parameters and calculates accumulated GDH for each of the supplied temperature time series.
Then it returns a dispersion metric measuring how different accumulated GDH are among the seasons.
The function is usually used in an optimization probelm, where the goal is to estimate
GDH model parameters most suitable to minimize the dispersion of accumulated GDH.

The function has very specific requirements on the structure of the Seasonlist.
In fact it is a list of Seasonlists created for several cultivars.
For each cultivar, the seasonlist contains exactly the period from endodormancy 
release to observed ecodormancy release.
}
\examples{
\dontrun{
#          yc,  zc,  s1, Tu,    E0,      E1,     A0,         A1,   Tf, Tc, Tb,  slope
par <-   c(40, 190, 0.5, 25, 3372.8,  9900.3, 6319.5, 5.939917e13,  4, 36,  4,  1.60)

#prepare weather data
weather<-fix_weather(KA_weather[which(KA_weather$Year>2004),])
hourtemps<-stack_hourly_temps(weather, latitude=50.4)
seasonList <- genSeasonList(hourtemps$hourtemps, year = 2006)

custom_PhenoFlex_GDHwrapper(x = seasonList[[1]], par = par)


}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
