% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_phenoflex_onlyreq.R
\name{eval_phenoflex_onlyreq}
\alias{eval_phenoflex_onlyreq}
\title{Evaluation function for combined fitting of cultivars
 
Takes model parameters for PhenoFlex, parameters of yc, zc, s1. Other parameters are fixed.
Used this function}
\usage{
eval_phenoflex_onlyreq(
  x,
  modelfn,
  bloomJDays,
  SeasonList,
  par_fixed,
  constrains = TRUE,
  return_pred = FALSE,
  na_penalty = 365
)
}
\arguments{
\item{x}{model parameters of PhenoFlex, with chill requirement (yc), 
heat requirement (zc), and transition parameter (s1) new format:}

\item{modelfn}{function used within the evaluation function to calculate the actual bloomday, often we use
the 'custom_GDH_wrapper' function for that}

\item{bloomJDays}{numeric containing the days of the year with the observed bloom}

\item{SeasonList}{list of hourly temperatures for the individual phonological seasons. Each element should contain a data.frame
with the columns "Temp" (for the hourly temperature) and "JDay" for the corresponding Julian day. Is usually
generated using \link[chillR]{genSeasonList}}

\item{par_fixed}{parameters of chill and heat submodel, order: Tu, E0, E1, A0, A1, Tf, Tc, Tb, slope}

\item{constrains}{logical, by default set TRUE. Indicates if there are further
checks on the parameters except the boundaries of the optimization algorithm}

\item{return_pred}{logical, by default FALSE. If set TRUE the function will return
bloom dates instead of the output required by the global optimzation algorithm.}

\item{na_penalty}{numeric, by default 365. Penalty for the phenology
prediction function when it fails return a bloom prediction}
}
\value{
list with two elements. First is called 'f' and contains the residual sum of squares of the model. The 
second is 'g' which is the values of the additional model constraints defined in the function.
If the flag for return_pred set TRUE, then it returns bloom dates
}
\description{
The function evaluates a set of parameters, it assumes the chill requirement (yc), heat
requirement (zc) and transition parameter (s1) cultivar-specific and the remaining
parameters for chill and heat submodels are shared. The function is mainly used
in model calibration and gets called by the global optimization algorithm.
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
