% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_phenoflex_single_twofixed.R
\name{eval_phenoflex_single_twofixed}
\alias{eval_phenoflex_single_twofixed}
\title{Function used in the optimization process to generate bloom data}
\usage{
eval_phenoflex_single_twofixed(
  x,
  modelfn,
  bloomJDays,
  SeasonList,
  Tc = 36,
  theta_star = 279,
  return_pred = FALSE,
  na_penalty = 365
)
}
\arguments{
\item{x}{traditional model parameters of PhenoFlex in the order yc, zc, s1, Tu, theta_c, tau, pie_c, Tf, Tb, slope.}

\item{modelfn}{function used within the evaluation function to calculate the actual bloomday, often we use
the 'custom_GDH_wrapper' function for that}

\item{bloomJDays}{numeric containing the days of the year witht the observed bloom}

\item{SeasonList}{list of hourly temperatures for the individual phenological seasons. Each element should contain a data.frame
with the columns "Temp" (for the hourly temperature) and "JDay" for the corresponding Julian day. Is usually
generated using \link[chillR]{genSeasonList}}

\item{Tc}{numeric, by default 36. Is the upper limit of temperature in degree Celsius, above which no heat accumulation happens.}

\item{theta_star}{numeric, by default 286. Is the optimal temperature for chill accumulation measured in Kelvin}

\item{return_pred}{boolean, by default FALSE. If set TRUE, then the evaluation function
does not return the performance score but the predicted bloom dates.}

\item{na_penalty}{numeric, value which is used when the model fails to generate a prediction
for the bloom date. By default 365}
}
\value{
list with two elements. First is called 'f' and contains the residual sum of squares of the model. The 
second is 'g' which is the values of the additional model constraints defined in the function.
}
\description{
This function works like \link[evalpheno]{evaluation_function_meigo} except that it uses 
parameters theta_star, theta_c, tau and pie_c instead of the commonly used
E0, E1, A0 and A1 parameters, which get calculated in an intermediate step.
}
\details{
This function follows the approach suggested in Egea 2021 and solves a 
nonlinear system of equations to derive the parameters E0, E1, A0 and A1 from the 
newly used parameters theta_star, theta_c, tau and pie_c. '
}
\author{
Lars Caspersen
}
\keyword{utility}
