% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalpheno-package.R
\docType{package}
\name{evalpheno-package}
\alias{evalpheno}
\alias{evalpheno-package}
\title{evalpheno: Custom Evaluation functions for PhenoFlex}
\description{
Contains evaluation functions and wrapper functions for calibrating the PhenoFlex model.
}
\author{
\strong{Maintainer}: Lars Caspersen \email{lcaspers@uni-bonn.de} (\href{https://orcid.org/0009-0000-3057-7327}{ORCID})

Authors:
\itemize{
  \item Lars Caspersen \email{lcaspers@uni-bonn.de} (\href{https://orcid.org/0009-0000-3057-7327}{ORCID})
}

}
\keyword{internal}
