% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_function_meigo.R
\name{evaluation_function_meigo}
\alias{evaluation_function_meigo}
\title{Function used in the optimization process to generate bloom data}
\usage{
evaluation_function_meigo(x, modelfn, bloomJDays, SeasonList, na_penalty = 365)
}
\arguments{
\item{x}{traditional model parameters of PhenoFlex in the order yc, zc, s1, Tu, E0, E1, A0, A1, Tf, Tc, Tb, slope.}

\item{modelfn}{function used within the evaluation function to calculate the actual bloomday, often we use
the 'custom_GDH_wrapper' function for that}

\item{bloomJDays}{numeric containing the days of the year witht the observed bloom}

\item{SeasonList}{list of hourly temperatures for the individual phenological seasons. Each element should contain a data.frame
with the columns "Temp" (for the hourly temperature) and "JDay" for the corresponding Julian day. Is usually
generated using \link[chillR]{genSeasonList}}

\item{na_penalty}{numeric, value which is used when the model fails to generate a prediction
for the bloom date. By default 365}
}
\value{
list with two elements. First is called 'f' and contains the residual sum of squares of the model. The 
second is 'g' which is the values of the additional model constraints defined in the function.
}
\description{
This function takes the temperature data, the phenological observations and model parameters and returns
a model performance score. This score is used by the optimization algorithm to 
find better model parameters.
}
\author{
Lars Caspersen
}
\keyword{utility}
