% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_function_meigo_nonlinear_fixed.R
\name{evaluation_function_meigo_nonlinear_fixed}
\alias{evaluation_function_meigo_nonlinear_fixed}
\title{Function used in the optimization process to generate bloom data}
\usage{
evaluation_function_meigo_nonlinear_fixed(
  x,
  modelfn,
  bloomJDays,
  SeasonList,
  na_penalty = 365,
  n_cult,
  theta_star = 286,
  pie_c = 28,
  Tc = 36,
  Tb = 2,
  slope = 1.6
)
}
\arguments{
\item{x}{traditional model parameters of PhenoFlex in the order yc, zc, s1, Tu, theta_star, theta_c, tau, pie_c, Tf, Tc, Tb, slope. The parameters
yc, zc and s1 need to be replicated as often as cultivars are fitted.}

\item{modelfn}{function used within the evaluation function to calculate the actual bloomday, often we use
the 'custom_GDH_wrapper' function for that}

\item{bloomJDays}{list of numeric vectors containing the observed bloom dates in day of the year format}

\item{SeasonList}{list of lists, containing the hourly temperatures for the individual phenological seasons. One element per cultivar.
The list of each cultivar should contain data.frames for eachs season
with the columns "Temp" (for the hourly temperature) and "JDay" for the corresponding Julian day. Is usually
generated using \link[chillR]{genSeasonList}}

\item{na_penalty}{numeric, value which is used when the model fails to generate a prediction
for the bloom date. By default 365}

\item{n_cult}{Numeric, specifying the numbers of cultivars fitted}

\item{theta_star}{numeric, by default 286. Is the optimal temperature for chill accumulation measured in Kelvin}

\item{pie_c}{numeric, ny default 28. Is the interval length in hours of a cycled two-temperature experiment which lead 
to absolute chill negation.}

\item{Tc}{numeric, by default 36. Is the upper limit of temperature in degree Celsius, above which no heat accumulation happens.}

\item{Tb}{numeric, by default 2. Is the lower limit of temperature in degree Celsius, below which no heat accumulation happens.}

\item{slope}{numeric, by default 1.6. Defines the slope of the transition process converting labile to stabile chill portions.}
}
\value{
list with two elements. First is called 'f' and contains the residual sum of squares of the model. The 
second is 'g' which is the values of the additional model constraints defined in the function.
}
\description{
This is the evaluation function for the fitting process. It takes the "new" parameters
(theta_star, theta_c, tau, pie_c) instead of the old ones (E0, E1, A0, A1). This function
allows combined fitting of cultivars. It will only estimate cultivar-specific parameters of yc, zc and s1. Furthermore
it has fixed values of theta_star, pie_c, Tc, Tu and slope.
}
\details{
This function will be best used in combination with the \link[MEIGOR]{MEIGO} function 
to estimate PhenoFlex parameter values. The order of bloomJDays and SeasonList should be the same. 
The order will correspond to estimated cultivar-specific parameters
}
\author{
Lars Caspersen
}
\keyword{utility}
