% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_function_meigo_vns.R
\name{evaluation_function_meigo_vns}
\alias{evaluation_function_meigo_vns}
\title{Function used in the optimization process to generate bloom data}
\usage{
evaluation_function_meigo_vns(
  x,
  modelfn,
  bloomJDays,
  SeasonList,
  na_penalty = 365
)
}
\arguments{
\item{x}{traditional model parameters of PhenoFlex in the order yc, zc, s1, Tu, E0, E1, A0, A1, Tf, Tc, Tb, slope.}

\item{modelfn}{function used within the evaluation function to calculate the actual bloomday, often we use
the 'custom_GDH_wrapper' function for that}

\item{bloomJDays}{numeric containing the days of the year witht the observed bloom}

\item{SeasonList}{list of hourly temperatures for the individual phenological seasons. Each element should contain a data.frame
with the columns "Temp" (for the hourly temperature) and "JDay" for the corresponding Julian day. Is usually
generated using \link[chillR]{genSeasonList}}

\item{na_penalty}{numeric, value which is used when the model fails to generate a prediction
for the bloom date. By default 365}
}
\value{
list with one element, which is called 'f' and contains the residual sum of squares of the model.
}
\description{
This function works like \code{\link{evaluation_function_meigo}} with some minor differences in
the model output to be compatible with the VNS optimizer of the MEIGO package.
}
\author{
Lars Caspersen
}
\keyword{utility}
