% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_model_list.R
\name{model_parameters}
\alias{model_parameters}
\title{Get stored parameters for calibration}
\usage{
model_parameters(model)
}
\arguments{
\item{model}{A single, population, combined or model-list specification, or a
result from \code{\link[=fit_phenology]{fit_phenology()}}.}
}
\value{
Complete named vector of stored calibration parameters. For model
lists this combines indexed member values and shared values.
}
\description{
Get stored parameters for calibration
}
