% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer_gensa_combined_fitting.R
\name{phenologyFitter_combined_fitting}
\alias{phenologyFitter_combined_fitting}
\title{Evaluation function for combined fitting of cultivars, using GenSA optimizer
 
Takes model parameters for PhenoFlex, parameters of yc, zc, s1 for individual cultivars and the remaining ones are shared. 
Used the function for the apple study in Bonn and Asturias (Hajar Mojahid)}
\usage{
phenologyFitter_combined_fitting(
  par.guess = NULL,
  modelfn = chillR::PhenoFlex_GDHwrapper,
  bloomJDays,
  SeasonList,
  n_cult,
  par_type = "old",
  control = list(smooth = FALSE, verbose = TRUE, maxit = 1000, nb.stop.improvement = 250),
  lower,
  upper,
  seed = 1235433,
  ...
)
}
\arguments{
\item{par.guess}{vector, numeric. Initial guess of the model parameters 
yc, zc, s1, Tu, E0, E1, A0, A1, Tf, Tc, Tb, slope}

\item{modelfn}{function used within the evaluation function to calculate the actual bloomday, by default we
are using the PhenoFlex_GDHwrapper f chillR package}

\item{bloomJDays}{numeric containing the days of the year with the observed bloom}

\item{SeasonList}{list of hourly temperatures for the individual phonological seasons. Each element should contain a data.frame
with the columns "Temp" (for the hourly temperature) and "JDay" for the corresponding Julian day. Is usually
generated using \link[chillR]{genSeasonList}}

\item{n_cult}{numeric, indicates the number of cultivars combined in the study}

\item{par_type}{character, controls if we use the old type (with E0, E1, A0, A1) or 
when equal to "new" for the format with (theta_star, theta_c, pie_c, tau)}

\item{control}{list, contains arguments controling the GenSA optimization algorithm.
For more details see in the documentation of the GenSA package.}

\item{lower}{vector, contains lower bound of the search space for the parameters}

\item{upper}{vector, contains upper bound of the search space for the parameters}

\item{seed}{numeric, sets seed to make the random process of optimization reproducible}

\item{...}{further arguments for the modelfn function}
}
\value{
list with two elements. First is called 'f' and contains the residual sum of squares of the model. The 
second is 'g' which is the values of the additional model constraints defined in the function.
If the flag for return_pred set TRUE, then it returns bloom dates
}
\description{
The function evaluates a set of parameters, it assumes the chill requirement (yc), heat
requirement (zc) and transition parameter (s1) cultivar-specific and the remaining
parameters for chill and heat submodels are shared. The function is mainly used
in model calibration and gets called by the global optimization algorithm.
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
