% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_JDay.R
\name{return_JDay}
\alias{return_JDay}
\title{Returns Julian Day (with fraction)}
\usage{
return_JDay(index, Jday_vec, year_vec = NULL)
}
\arguments{
\item{index}{integer, when heat requirement is met}

\item{Jday_vec}{vector of julian days}

\item{year_vec}{vector with the year}
}
\value{
day of the year, with a fraction to account for when (at which hour)
the requirement is met. 12:00 midday corresponds to .0, in the afternoon up to +0.5 is added to the julian day, 
if it is before midday up to -0.5 is added to the Julian Day
}
\description{
Function needed to handle the calibration function for several phenology stages
}
\examples{
\dontrun{
 
 i <- 56
 Jday_vec <- rep(1:30, each = 24)
 year_vec <- rep(2024, length(Jday_vec))
 return_JDay(i, Jday_vec, year_vec)
}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
