% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_parallel_model.R
\name{wrapper_parallel_model}
\alias{wrapper_parallel_model}
\title{Wrapper for Parallel Model}
\usage{
wrapper_parallel_model(x, par)
}
\arguments{
\item{x}{data.frame with at least the column "Temp" (hourly temperature data) and "JDay" (day of the year)}

\item{par}{vector of length 12 with the parameters of the Parallel model (3), chill submodel (6) and heat submodel (3):
parameters from parallel model include: yc (chill requirement), zc (heat requirement), kmin (minimum relative heat effectiveness at zero chill)
parameters of chill submodel (Dynamic Model) include: E0, E1, A0, A1, Tf, slope. See \link[chillR]{Dynamic_Model} for more information.
parameters of heat submodel (Growing Degree Hour model): Tb, Tu, Tc. See \link[chillR]{GDH} for more information.}
}
\value{
single numeric value with the day of the year, for which the model predicts bloom
with given temperature data and model parameters.
}
\description{
Takes hourly weather time series and parameters from Parallel model and returns day of the year with bloom.
}
\details{
This function is mainly used by function called by the optimizer during calibration
}
\references{
Hänninen, H. and Kramer, K. (2007). A framework for modelling the annual cycle of trees in boreal and temperate regions. Silva Fennica 41(1), 167-205. Equation B4b. doi:10.14214/sf.313.
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
