% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_seq_model.R
\name{wrapper_seq_model}
\alias{wrapper_seq_model}
\title{Wrapper for Sequential Model}
\usage{
wrapper_seq_model(x, par)
}
\arguments{
\item{x}{data.frame with at least the column "Temp" (hourly temperature data) and "JDay" (day of the year)}

\item{par}{vector of length 11 with the parameters of the Sequential model (2), chill submodel (6) and heat submodel (3):
parameter of sequenital model include: yc (chill requirement), zc (heat requirement).
parameters of chill submodel (Dynamic Model) include: E0, E1, A0, A1, Tf, slope. See \link[chillR]{Dynamic_Model} for more information.
parameters of heat submodel (Growing Degree Hour model): Tb, Tu, Tc. See \link[chillR]{GDH} for more information.}
}
\value{
single numeric value with the day of the year, for which the model predicts bloom
with given temperature data and model parameters.
}
\description{
Takes hourly weather time series and parameters from Sequential model and returns day of the year with bloom.
}
\details{
This function is mainly used by function called by the optimizer during calibration
}
\references{
Ashcroft, G. L., Richardson, E. A., & Seeley, S. D. (1977). A statistical method of determining chill unit and growing degree hour requirements for deciduous fruit Trees1. HortScience, 12(4), 347-348.
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
