% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{po_model}
\alias{po_model}
\title{partial_overlap_model}
\usage{
po_model(
  temp,
  times,
  yc = 40,
  b1 = 1119,
  b2 = 8677,
  b3 = 0.01119,
  ol = 0.75,
  A0 = 6319.5,
  A1 = 5.939917e+13,
  E0 = 3372.8,
  E1 = 9900.3,
  slope = 1.6,
  Tf = 4,
  Tu = 25,
  Tb = 4,
  Tc = 36,
  Delta = 4,
  stopatzc = TRUE,
  deg_celsius = TRUE,
  basic_output = TRUE
)
}
\arguments{
\item{yc}{numeric, Critical value defining end of chill accumulation}

\item{b1}{numeric. Heat requirement at minimum chilling}

\item{b2}{numeric. Heat requiement at maximum chilling}

\item{b3}{numeric. Scales the compensation from heat and chill requirement. Low b3 leads to linear compensation, high b3 to much heat needed to compensate chill.}

\item{ol}{numeric. Controls how much longer chill accumulates after reaching yc. Chill accumulation stops when share of heat requirement is surprassed.}

\item{A0}{numeric. Parameter \eqn{A_0}{A0} of the dynamic model}

\item{A1}{numeric. Parameter \eqn{A_1}{A1} of the dynamic model}

\item{E0}{numeric. Parameter \eqn{E_0}{E0} of the dynamic model}

\item{E1}{numeric. Parameter \eqn{E_1}{E1} of the dynamic model}

\item{slope}{numeric. Slope parameter for sigmoidal function //' @param Tf numeric. Transition temperature (in degree Kelvin) for the sigmoidal function}

\item{Tu}{numeric. GDH optimal temperature}

\item{Tb}{numeric. GDH base temperature (lower threshold)}

\item{Tc}{numeric. GDH critical temperature}

\item{Delta}{numeric. Width of Gaussian heat accumulation model}

\item{stopatzc}{boolean. If `TRUE`, the PhenoFlex is applied until the end of the temperature series. Default is to stop once the value zc has been reached.}

\item{basic_output}{boolean. If `TRUE`, only the bloomindex is returned as a named element of the return list.}
}
\value{
A list is returned with named element `bloomindex`, which is the index at which blooming occurs. When `basic_output=FALSE` also `x`, `y`, `z` and `xs` are
returned as named element of this list, which are numeric vectors of the same length as the input vector `temp` containing the hourly temperatures.
}
\description{
Partial Overlap Model, combining the dynamic model for chill accumulation and the GDH model
}
\examples{
data(KA_weather)
hourtemps <- stack_hourly_temps(KA_weather, latitude=50.4)
iSeason <- genSeason(hourtemps, years=c(2009))
yc <- 40
b1 <- 1119
b2 <- 8677
b3 <- 0.01119
ol <- 0.75
x <- po_model(temp=hourtemps$hourtemps$Temp[iSeason[[1]]],
               times=c(1: length(hourtemps$hourtemps$Temp[iSeason[[1]]])),
               yc=yc, b1 = b1, b2=b2, b3=b3, ol = ol, stopatzc=TRUE, basic_output=FALSE)
DBreakDay <- x$bloomindex
ii <- c(1:DBreakDay)
plot(x=ii, y=x$z[ii], xlab="Hour Index", ylab="z", col="red", type="l")
abline(h=zc, lty=2)
plot(x=ii, y=x$y[ii], xlab="Hour Index", ylab="y", col="red", type="l")
abline(h=yc, lty=2)
}
\author{
Lars Caspersen <lcaspers@uni-bonn.de>
}
