% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_po_model.R
\name{wrapper_po_model}
\alias{wrapper_po_model}
\title{Wrapper for Partial Overlap Model}
\usage{
wrapper_po_model(x, par)
}
\arguments{
\item{x}{data.frame with at least the column "Temp" (hourly temperature data) and "JDay" (day of the year)}

\item{par}{vector of length 14 with the parameters of the Partial Overlap model (5), chill submodel (6) and heat submodel (3):
yc (chill requirement), b1 (heat requirement at marginal chilling), b2 (heat requirement at optimal chilling), b3 (steepness of chill and heat compensation (low b3 = linear compensation, high b3 = much more heat needed at marginal chill), ol = extra chill accumulation after reachining minimum chill requirement, up to overlap * heat requirement.
parameters of chill submodel (Dynamic Model) include: E0, E1, A0, A1, Tf, slope. See \link[chillR]{Dynamic_Model} for more information.
parameters of heat submodel (Growing Degree Hour model): Tb, Tu, Tc. See \link[chillR]{GDH} for more information.}
}
\value{
single numeric value with the day of the year, for which the model predicts bloom
with given temperature data and model parameters.
}
\description{
Takes hourly weather time series and parameters from Partial Overlap model and returns day of the year with bloom.
}
\details{
This function is mainly used by function called by the optimizer during calibration
}
\references{
Pope, K.S., Da Silva, D., Brown, P.H., DeJong, T.M. (2014). A biologically based approach to modeling spring phenology in temperate deciduous trees. Agricultural and Forest Meteorology, 198, 15-23. https://doi.org/10.1016/j.agrformet.2014.07.009
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
